+++ Libraries/LibWebView/EventLoop/EventLoopImplementationMacOS.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Tim Flynn <trflynn89@serenityos.org>
+ * Copyright (c) 2023-2024, Tim Flynn <trflynn89@ladybird.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -10,9 +10,9 @@
 #include <AK/NonnullOwnPtr.h>
 #include <LibCore/EventLoopImplementation.h>
 
-namespace Ladybird {
+namespace WebView {
 
-class CFEventLoopManager final : public Core::EventLoopManager {
+class EventLoopManagerMacOS final : public Core::EventLoopManager {
 public:
     virtual NonnullOwnPtr<Core::EventLoopImplementation> make_implementation() override;
 
@@ -28,12 +28,12 @@ public:
     virtual void unregister_signal(int) override;
 };
 
-class CFEventLoopImplementation final : public Core::EventLoopImplementation {
+class EventLoopImplementationMacOS final : public Core::EventLoopImplementation {
 public:
     // FIXME: This currently only manages the main NSApp event loop, as that is all we currently
     //        interact with. When we need multiple event loops, or an event loop that isn't the
     //        NSApp loop, we will need to create our own CFRunLoop.
-    static NonnullOwnPtr<CFEventLoopImplementation> create();
+    static NonnullOwnPtr<EventLoopImplementationMacOS> create();
 
     virtual int exec() override;
     virtual size_t pump(PumpMode) override;
@@ -47,7 +47,7 @@ public:
     virtual void notify_forked_and_in_child() override { }
 
 private:
-    CFEventLoopImplementation() = default;
+    EventLoopImplementationMacOS() = default;
 
     int m_exit_code { 0 };
 };
