+++ Services/WebWorker/main.cpp
@@ -25,8 +25,8 @@
 #include <WebWorker/ConnectionFromClient.h>
 
 #if defined(HAVE_QT)
+#    include <LibWebView/EventLoop/EventLoopImplementationQt.h>
 #    include <QCoreApplication>
-#    include <UI/Qt/EventLoopImplementationQt.h>
 #endif
 
 static ErrorOr<void> initialize_resource_loader(JS::Heap&, int request_server_socket);
@@ -52,7 +52,7 @@ ErrorOr<int> serenity_main(Main::Arguments arguments)
 
 #if defined(HAVE_QT)
     QCoreApplication app(arguments.argc, arguments.argv);
-    Core::EventLoopManager::install(*new Ladybird::EventLoopManagerQt);
+    Core::EventLoopManager::install(*new WebView::EventLoopManagerQt);
 #endif
     Core::EventLoop event_loop;
